Updated README.md with Ctrl bindings only
This is the way I expected this package.
As an ex-Atom user on Linux, there was the Ctrl-* for normal word navigation, like every other editors and programs, and then the added Alt-* for subword navigation. And I like it pretty much like this. Emacs may use it this way too.

Your keybinding.json snippet for OSX should be the default on I think.